Deposit refactor
