+++ contracts/carrot-app/src/distribution/mod.rs
@@ -1,0 +1,4 @@
+/// This module handles the strategy distribution. It handles the following cases
+/// 1. A user deposits some funds.
+/// This modules dispatches the funds into the different strategies according to current status and target strats
+pub mod deposit;
