+++ contracts/carrot-app/src/handlers/instantiate.rs
@@ -28,7 +28,6 @@ pub fn instantiate_handler(
         .rewards
         .check(deps.as_ref(), &msg.dex, ans.host())?;
 
-    let strategy_len = msg.balance_strategy.0.len();
     let config: Config = Config {
         dex: msg.dex,
         balance_strategy: msg.balance_strategy,
@@ -40,8 +39,7 @@ pub fn instantiate_handler(
 
     // If provided - do an initial deposit
     if let Some(funds) = msg.deposit {
-        let deposit_msgs =
-            _inner_deposit(deps.as_ref(), &env, funds, vec![None; strategy_len], &app)?;
+        let deposit_msgs = _inner_deposit(deps.as_ref(), &env, funds, None, &app)?;
 
         response = response.add_messages(deposit_msgs);
     }
